# test: Speed up worker tests