Made some massive performance improvements to the worker test:

Old top 8 times:
```
141.053s   test_update_bucket_cve (worker_test.UpdateTest.test_update_bucket_cve)
73.236s    test_update_redhat_toobig (worker_test.RESTUpdateTest.test_update_redhat_toobig)
72.607s    test_update_alpine (worker_test.UpdateTest.test_update_alpine)                                    ->   50.822s
22.779s    test_analysis_crash_handling (worker_test.UpdateTest.test_analysis_crash_handling)
10.354s    test_update_conflict_while_pushing (worker_test.UpdateTest.test_update_conflict_while_pushing)
7.400s     test_get_ecosystem (worker_test.EcosystemTest.test_get_ecosystem)
3.205s     test_update_debian (worker_test.UpdateTest.test_update_debian)
2.135s     test_update_bucket (worker_test.UpdateTest.test_update_bucket)
```

New top 8 times:

```
53.291s    test_update_alpine (worker_test.UpdateTest.test_update_alpine)
10.322s    test_update_conflict_while_pushing (worker_test.UpdateTest.test_update_conflict_while_pushing)
7.654s     test_get_ecosystem (worker_test.EcosystemTest.test_get_ecosystem)
4.275s     test_update_bucket_cve (worker_test.UpdateTest.test_update_bucket_cve)
2.604s     test_update_debian (worker_test.UpdateTest.test_update_debian)
1.626s     test_update_android (worker_test.UpdateTest.test_update_android)
1.522s     test_update_bucket (worker_test.UpdateTest.test_update_bucket)
1.234s     test_update_last_affected (worker_test.UpdateTest.test_update_last_affected)
```
---

Lists of things done:
- Removed `test_analysis_crash_handling`, as it was no longer testing what the initial PR described it as testing:  https://github.com/google/osv.dev/pull/2884
  - The original PR says that this was to test the behavior of git ranges causing errors, but this test wasn't actually hitting any of the code changed in that PR. :shrug:
  - It also took 22 seconds.
- Skipped the redhat_toobig test, since it takes almost 80s locally, to essentially test that datastore will reject the entry because it is too big, which isn't that helpful to know or confirm constantly.
- Change the bucket worker test to use a smaller repo, reduced from 130s locally to ~1.5s
- Added some "caching" logic to repo reset function, so that if the repo has been reset recently (5 mins), don't fetch and reset it again, this speeds up Alpine, which switches branches a lot to cover each of the alpine releases.
   - 75s --> 63s

---

Fixes: #3893